api: use LocalProxy instead of decorator
* Introduces single variable local proxy instead of decorator- before we
depended on the with_api_client decorator, and the click context- now it's only
one LocalProxy.
* Removes Client class because it was using inheritance (from BaseAPIClient)
and it wasn't a correct relation. Without it it was just a collection of static
methods so we decided to leave them as separate functions.
* Removes mock_base_api_client fixture because it was returning a Client class
instance, when the class is no longer present. Something similar is
necessery because now we have a repeating code in lots of places-see the
issue: https://github.com/reanahub/reana-client/issues/219#issue-388666174