cli: improve reana spec handling in `download` and `upload` commands
After introducing missing workflow spec schema to `get_workflow_specification`
endpoint, bravado started to include and set all the missing properties
to `None` in the response of this endpoint. This affected `upload` and `download`
commands since internally it is used to get the files for upload/download.
This commit fixes the handling of these additional properties.